+++ objc2/src/message/mod.rs
@@ -1,32 +1,43 @@
-use alloc::string::String;
-use core::fmt;
 use core::mem;
 use core::mem::ManuallyDrop;
 use core::ptr::NonNull;
-use std::error::Error;
 
 use crate::rc::{Id, Owned, Ownership};
 use crate::runtime::{Class, Imp, Object, Sel};
-#[cfg(feature = "malloc")]
-use crate::verify::VerificationError;
 use crate::{Encode, EncodeArguments, RefEncode};
 
 #[cfg(feature = "catch_all")]
-unsafe fn conditional_try<R: Encode>(f: impl FnOnce() -> R) -> Result<R, MessageError> {
-    use alloc::borrow::ToOwned;
-    unsafe { crate::exception::catch(f) }.map_err(|exception| {
-        if let Some(exception) = exception {
-            MessageError(alloc::format!("Uncaught exception {:?}", exception))
-        } else {
-            MessageError("Uncaught exception nil".to_owned())
+#[track_caller]
+unsafe fn conditional_try<R: Encode>(f: impl FnOnce() -> R) -> R {
+    match unsafe { crate::exception::catch(f) } {
+        Ok(r) => r,
+        Err(exception) => {
+            if let Some(exception) = exception {
+                panic!("uncaught exception {:?}", exception)
+            } else {
+                panic!("uncaught exception nil")
+            }
         }
-    })
+    }
 }
 
 #[cfg(not(feature = "catch_all"))]
 #[inline]
-unsafe fn conditional_try<R: Encode>(f: impl FnOnce() -> R) -> Result<R, MessageError> {
-    Ok(f())
+#[track_caller]
+unsafe fn conditional_try<R: Encode>(f: impl FnOnce() -> R) -> R {
+    f()
+}
+
+#[cfg(feature = "verify_message")]
+#[track_caller]
+fn panic_verify(cls: &Class, sel: Sel, err: crate::VerificationError) -> ! {
+    panic!(
+        "invalid message send to {}[{:?} {:?}]: {}",
+        if cls.is_metaclass() { "+" } else { "-" },
+        cls,
+        sel,
+        err
+    )
 }
 
 #[cfg(feature = "apple")]
@@ -150,7 +161,8 @@ pub unsafe trait MessageReceiver: private::Sealed + Sized {
     /// The added invariant is that the selector must take the same number of
     /// arguments as is given.
     #[inline]
-    unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> Result<R, MessageError>
+    #[track_caller]
+    unsafe fn send_message<A, R>(self, sel: Sel, args: A) -> R
     where
         A: MessageArguments,
         R: Encode,
@@ -164,10 +176,12 @@ pub unsafe trait MessageReceiver: private::Sealed + Sized {
             let cls = if let Some(this) = this {
                 this.class()
             } else {
-                return Err(VerificationError::NilReceiver(sel).into());
+                panic!("messsaging {:?} to nil", sel);
             };
 
-            cls.verify_sel::<A, R>(sel)?;
+            if let Err(err) = cls.verify_sel::<A, R>(sel) {
+                panic_verify(cls, sel, err);
+            }
         }
         unsafe { send_unverified(this, sel, args) }
     }
@@ -192,12 +206,8 @@ pub unsafe trait MessageReceiver: private::Sealed + Sized {
     /// The added invariant is that the selector must take the same number of
     /// arguments as is given.
     #[inline]
-    unsafe fn send_super_message<A, R>(
-        self,
-        superclass: &Class,
-        sel: Sel,
-        args: A,
-    ) -> Result<R, MessageError>
+    #[track_caller]
+    unsafe fn send_super_message<A, R>(self, superclass: &Class, sel: Sel, args: A) -> R
     where
         A: MessageArguments,
         R: Encode,
@@ -206,9 +216,11 @@ pub unsafe trait MessageReceiver: private::Sealed + Sized {
         #[cfg(feature = "verify_message")]
         {
             if this.is_null() {
-                return Err(VerificationError::NilReceiver(sel).into());
+                panic!("messsaging {:?} to nil", sel);
             }
-            superclass.verify_sel::<A, R>(sel)?;
+            if let Err(err) = superclass.verify_sel::<A, R>(sel) {
+                panic_verify(superclass, sel, err);
+            }
         }
         unsafe { send_super_unverified(this, superclass, sel, args) }
     }
@@ -374,38 +386,6 @@ message_args_impl!(
     l: L
 );
 
-/// An error encountered while attempting to send a message.
-///
-/// Currently, an error may be returned in two cases:
-///
-/// - an Objective-C exception is thrown and the `catch_all` feature is
-///   enabled
-/// - the encodings of the arguments do not match the encoding of the method
-///   and the `verify_message` feature is enabled
-// Currently not Clone for future compat
-#[derive(Debug, PartialEq, Eq, Hash)]
-pub struct MessageError(String);
-
-impl fmt::Display for MessageError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Display::fmt(&self.0, f)
-    }
-}
-
-impl Error for MessageError {
-    fn description(&self) -> &str {
-        &self.0
-    }
-}
-
-#[cfg(feature = "malloc")]
-impl<'a> From<VerificationError<'a>> for MessageError {
-    fn from(err: VerificationError<'_>) -> MessageError {
-        use alloc::string::ToString;
-        MessageError(err.to_string())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -453,8 +433,11 @@ mod tests {
         assert_eq!(result, expected);
     }
 
-    #[cfg(not(feature = "verify_message"))]
     #[test]
+    #[cfg_attr(
+        feature = "verify_message",
+        should_panic = "messsaging description to nil"
+    )]
     fn test_send_message_nil() {
         use crate::rc::Shared;