+++ tests/assembly/test_msg_send_zero_cost/lib.rs
@@ -7,7 +7,7 @@ use objc2::MessageReceiver;
 
 #[no_mangle]
 unsafe fn handle(obj: &Object, sel: Sel) -> *mut Object {
-    MessageReceiver::send_message(obj, sel, ()).unwrap()
+    MessageReceiver::send_message(obj, sel, ())
 }
 
 // This will definitely not work, but is useful for making the assembly look
@@ -23,5 +23,5 @@ fn selector() -> Sel {
 
 #[no_mangle]
 unsafe fn handle_with_sel(obj: &Object) -> *mut Object {
-    MessageReceiver::send_message(obj, selector(), ()).unwrap()
+    MessageReceiver::send_message(obj, selector(), ())
 }